requirements.txt
pygame==2.5.2
numpy==1.24.3
numba==0.58.1
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
glfw==2.6.2
